+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.1
+* Now you can use custom transitions between pages.
+* By default the transition is Cupertino on macOS/iOS and Material on android.
+
 ## 0.1.0
 
 * **Breaking changes**